# Used datetime.timezone.utc instead of pytz.utc for better performance.