Before:
```
In [2]: %timeit django.utils.timezone.now()
2.77 µs ± 57.8 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```
After:
```
In [2]: %timeit django.utils.timezone.now()
937 ns ± 5.35 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```