@blueyed 2018-06-21
Can `get_default_timezone` be made to return the same for UTC?

Currently a test in DRF is failing with dj21 because of this.

It baseically does this:

```python
In [1]: from django.conf import settings
In [2]: settings.configure(USE_TZ=True, TIME_ZONE='UTC')
In [3]: from django.utils.timezone import get_default_timezone, utc
In [4]: get_default_timezone(), utc
Out[4]: (<UTC>, datetime.timezone(datetime.timedelta(0), 'UTC'))
In [5]: get_default_timezone() == utc
Out[5]: False
```

The test can be fixed of course, but that might affect other apps, and if it is meant for performance should be done there then, too.